Move addFunction/removeFunction to JS library
Depends on #17369 which enabled these functions to continue to be
exported via `EXPORTED_RUNTIME_METHODS`, and ensure good error messages
informing users how to include these library functions.

Hopefully many more changes like this to come but this is really testing
the waters.
