+++ tests/other/metadce/test_metadce_hello_O0.jssize
@@ -1,1 +1,1 @@
-26618
+26811
