+++ tests/other/metadce/test_metadce_minimal_O0.jssize
@@ -1,1 +1,1 @@
-21309
+21502
