+++ tests/other/test_unoptimized_code_size.js.size
@@ -1,1 +1,1 @@
-98207
+93804
