+++ tests/other/test_unoptimized_code_size_no_asserts.js.size
@@ -1,1 +1,1 @@
-59667
+54755
